[IDE] Move ExprContextAnalyser to its own file
NFC
