+++ lib/IDE/CMakeLists.txt
@@ -2,6 +2,7 @@ add_swift_host_library(swiftIDE STATIC
   CodeCompletion.cpp
   CodeCompletionCache.cpp
   CommentConversion.cpp
+  ExprContextAnalysis.cpp
   Formatting.cpp
   Refactoring.cpp
   ModuleInterfacePrinting.cpp
@@ -14,6 +15,7 @@ add_swift_host_library(swiftIDE STATIC
   APIDigesterData.cpp
   SourceEntityWalker.cpp
   LINK_LIBRARIES
+    swiftAST
     swiftFrontend
     swiftClangImporter
     swiftDWARFImporter
